pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "composearr"
version = "0.1.0"
description = "Docker Compose hygiene linter with cross-file intelligence"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "coaxk" },
]
keywords = ["docker", "compose", "linter", "hygiene", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer[all]>=0.12.0",
    "ruamel.yaml>=0.18.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
    "inquirerpy>=0.3.4",
]

[project.scripts]
composearr = "composearr.cli:app"

[project.urls]
Homepage = "https://github.com/coaxk/composearr"
Repository = "https://github.com/coaxk/composearr"

[tool.hatch.build.targets.wheel]
packages = ["src/composearr"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
network = [
    "requests>=2.31.0,<3.0.0",
    "packaging>=24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
